First Version!
This was the first prototype of my idea for this app, which I made back in June of 2021. This repo and commit are actually being started on early Valentine's Day of 2022 (Feb 14th), as now I have more of an understanding with async/await and the File System Access API!
I have started looking into ES Modules for the web in the last week or so, and that has been my new learning curve recently. I started getting into object classes with Flatlands, so I want to see what I can make with all of these new ideas and see how each of them may concide inside a new PWA project!

One of my main goals for starting this project is to have a proper, simple music player for my iTunes Library file system on Windows. I use Bootcamp on my Mac, and I have it set up so both operating systems can access the library song files from either partition, but the Windows version of iTunes doesn't support the new folder naming format that the new "Music" app does with macOS (The one that replaced iTunes in Catalina).
Using the File System Access API, I can just make my own PWA file system directory tree-viewer with enhanced music playing capabilities!
I also just remembered/came across a way to save File System Handles to IndexedDB (also new to me, still using LocalStorage at the moment, haha), which is very futuristic! This means the app will only have to prompt me to access the folder on load, no need to reselect it again to open my library or anything.

Can't wait to see what comes out of this project!
And I just came up with the current name "Librar.io" today, so that could possibly change, but we'll have to see!
Oh yeah, and in my mind, it's pronounced either "Li-brar-eeoo", like a music library, but with the `.io` naming convention a lot of sites are using, because why not! It's more of a meta-humor reference to that than anything else, haha.