JSMediaTags Module + Unhandled Error Caching
Added the JSMediaTags module! This is a really nice library I found that will get song attributes out of most audio file formats. I will be able to access the album artwork saved in the file, and all of the song metadata, which will be very helpful for making the app UI accurately represent the song's info.