Drag and Drop Rewrite!
Wow! That took some big brainpower. I'll just show you this explainer paragraph I wrote about what it's doing, I can't write it again! XD

The block of code on Line 17 essentially does this:

1. Get the items from the DataTransfer items *object*
2. Turn that into a standard array, so then you can use array methods to manage the content.
3. Filter that new array to only include DataTransfer items that are Files, not strings or other possible DataTransfer types (Files in this step mean items from the File System, so eventually either a real File or Directory will match here).
4. Create a new array from information about those Files, namely, replace each DataTransferItem object with it's corresponding FileSystemHandle object result.
5. Wait for all of these FileSystemHandle objects in the array to resolve, the return those.
6. Now filter over each of these and determine if it is a real File or a Directory, in terms of the File System Access API.

Done! Now `directoryHandles` is an array of the FileSystemDirectoryHandle objects of the folders that you dropped on the document. Amazing!!!

I will definitely make this into it's own ES Module, I just wanted to commit this to free up my brain space again, this was monsterous!