Artwork Collection Fix
Haven't found a consistent way to garbage collect the Blob URLs for album art, as sometimes it errors out when trying to access it.
Without this line removed, once you started playing the song, the Blob URL had already been revoked, so the Media Session API couldn't get to it in time. I think I need to store a value that holds the current song info, only while that song is playing, then revoke it's URLs and remove it's metadata once you play the next song. This would make it so only one song is in memory at a given time, including it's metadata and it's album art. I think not taking care of the loaded album art could lead to memory leaks if not taken care of correctly.