TSC Time! Part Two
Wow! That took some big brain magic to get fully typed-out. Found some wonky/missing type definitions for the File System Access API, and that kind of stumped me on a few problems for a while. But, in the end, I got it all done tonight!

This update simply adds type checking to the existing codebase, and adds some of the dev dependencies required to make that fully work. I also had to make a few changes to the code itself in a few spots, just to make the type definitions work like they should.

Learned a lot about my existing codebase from this! It's amazing what things it points out to you, considering I know that I built this, but am still learning how it works yet again XD

Had to use `// @ts-expect-error` in a few spots. Didn't know too much about that one, vs simply just `// @ts-ignore`, so I decided to look into it to see if I should use that one instead. This little article was neat for that :)
https://dev.to/oliverjumpertz/making-good-use-of-ts-expect-error-in-typescript-1f41

Also, found out how to semi-consistently get drag-and-drop to work on Ubuntu! It only works from the desktop (not Nautilus), and you also have to re-focus Chrome before you re-hover over it to place the item there. If you do both of those things, it works about 65% of the time, versus 0% XD

It looks like there's some type issues with the `new-javascript` type definitions, so the project isn't fully `tsc` compliant yet. None of my files have any issues within the editor though! That's much closer than before.

Merry Christmas!
Not sure why I'm up so late in the morning, working on this. I should go to bed! My Chromebook is at 5% too, so I'm squeezing this one in just in time, haha.

...and to all a goodnight! :)