Directory Iteration
Found a type definition difference in the `new-javascript` library that I'm using for TypeScript definitions for the File System Access API, so I made a suggestion to add the types for it. Big thanks to BenjaminAster for the quick response and fix to add the type definitions for it!
https://github.com/BenjaminAster/TypeScript-types-for-new-JavaScript/issues/1

I did find a few other issues with the types though, in some spots it appears to clash with the default TypeScript library. I'm going to see if I can fix them myself, then submit a PR to fix it too, as it appears to be `class` definitions vs `interface` ones. I think the TypeScript standard library depends on the use of interface merging to not clash with doubly-defined type definitions, and `new-javascript` appears to use `class` definitions in some spots. I originally didn't know the difference between these two different ways to do it, and now it makes more sense why libraries don't use `class` definitions as much, since they can't appear in multiple spots, because they can't be redefined, unline `interface` and `declare var` definitions, which seem like they can be, at least to my knowledge. Gonna try the PR fix, and see if that solves it! If it does, then I think I'm learning on the right track!