New-JavaScript Type Errors: Resolved!
Another epic thanks to BenjaminAster! Turns out the type definitions were clashing with the standard library because some of these bleeding-edge types were officially added, so the ones defined in `new-javascript` were being applied on top of the new ones recently added to `typescript`. It also had to do with some references to the `WebWorker` definitions, which were clashing with the ones defined in the `DOM` definitions.
https://github.com/BenjaminAster/TypeScript-types-for-new-JavaScript/issues/2